src/services/query_service/pyproject.toml
# services/query-service/pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "query-service"
version = "0.1.0"
description = "Portfolio Analytics - API Query Service"
requires-python = ">=3.11"
dependencies = [
    "fastapi==0.111.0",
    "uvicorn[standard]==0.30.1",
    "sqlalchemy==2.0.30",
    "asyncpg==0.29.0",
    "pydantic==2.7.4",
    "python-dotenv==1.0.1",
    "prometheus-fastapi-instrumentator==7.0.0",
    "python-dateutil>=2.8"
]

[tool.setuptools.packages.find]
where = ["app"]